Release 0.1.1

+++ MPM-Betting.DataModel/User/MpmUser.cs
@@ -10,7 +10,7 @@ public class MpmUser : IdentityUser
 {
     public List<UserGroupEntry> UserGroupEntries { get; set; } = [];
     
-    public int Points { get; set; } = 0;
+    public long Points { get; set; } = 0;
     public DateTime LastRedeemed { get; set; } = DateTime.MinValue;
     
     public List<Achievement> Achievments { get; set; } = [];
